requirements.txt
fastapi==0.128.0
httpx==0.28.1
python-dotenv==1.2.1
uvicorn==0.40.0
